# PERF: TimedeltaArray.astype(object)