```
In [2]: tdi = pd.timedelta_range("1 second", periods=10**5, freq="s")

In [5]: %timeit tdi.astype(object)
388 ms ± 43.5 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   # <- PR
588 ms ± 63.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)   # <- main
```